Refactor parser files to a new .parser package
